[SHT2x] Fix integer type for mesaurments [SHT2x] Fix wrong assembly of serial number.

+++ platform/humsen/sht2x.h
@@ -31,8 +31,8 @@ typedef struct sht2x* sht2x_t;
 sht2x_t sht2x_create(i2c_dev_t bus);
 
 // in mili celsius
-int16_t sht2x_measure_temperature(sht2x_t dev);
+int32_t sht2x_measure_temperature(sht2x_t dev);
 // in %RH/1000
-int16_t sht2x_measure_humidity(sht2x_t dev);
+int32_t sht2x_measure_humidity(sht2x_t dev);
 
 uint64_t sht2x_read_serial_number(sht2x_t dev);
